Moved and updated the header files
