+++ src/memory.cc
@@ -25,7 +25,7 @@
 //
 // =================================================================================================
 
-#include "tuner/internal/memory.h"
+#include "internal/memory.h"
 
 namespace cltune {
 // =================================================================================================
