+++ src/searcher.cc
@@ -25,7 +25,7 @@
 //
 // =================================================================================================
 
-#include "tuner/internal/searcher.h"
+#include "internal/searcher.h"
 
 #include <limits>
