+++ src/searchers/full_search.cc
@@ -25,7 +25,7 @@
 //
 // =================================================================================================
 
-#include "tuner/internal/searchers/full_search.h"
+#include "internal/searchers/full_search.h"
 
 namespace cltune {
 // =================================================================================================
